Roll prod.rs into lib.rs in ykshim and ykshim_client.
No functional change.
