+++ internal_ws/ykshim/src/test_api.rs
@@ -1,3 +1,7 @@
+//! The ykshim testing API.
+//!
+//! These functions are only exposed to allow testing from the external workspace.
+
 use libc::size_t;
 use std::convert::TryFrom;
 use std::default::Default;
